Demo-App-Flask-mvc/requirements.txt
certifi==2021.10.8        
certifi==2021.10.8        
certifi==2021.10.8        
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
dnspython==2.2.1
dominate==2.6.0
email-validator==1.1.3
Flask==2.1.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.0
Flask-Migrate==3.1.0
flask-mongoengine==1.0.0
flask-oidc==1.4.0
Flask-PyMongo==2.3.0
Flask-SQLAlchemy==2.5.1
flask-swagger==0.2.14
flask-swagger-ui==3.36.0
Flask-WTF==1.0.1
greenlet==1.1.2
httplib2==0.20.4
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.7.1
itsdangerous==2.0.1
Jinja2==3.1.1
Mako==1.2.0
MarkupSafe==2.1.1
mongoengine==0.24.1
oauth2client==4.1.3
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pymongo==4.1.1
pyparsing==3.0.8
PyYAML==6.0
requests==2.27.1
rsa==4.8
six==1.16.0
SQLAlchemy==1.4.34
urllib3==1.26.9
visitor==0.1.3
Werkzeug==2.1.1
WTForms==3.0.1
zipp==3.8.0
drf-yasg